Create .py
I created a programme that tells you how many days there are in every given month for any given year. This program will tell you how many days are in that given month including if that year turns out to be a leap year, which means that instead of 365 days, we'll have 366 days. It will also tell you if February has 28 or 29 days, in that year.